refactor: Database queries optimized and types adjusted
- Adjusted `Episode.Runtime`, `SeasonNumber`, `EpisodeNumber` to int32.
- Changed movie and tv show runtime to int32.
- Converted user IDs to int64 in database queries.
- Watchlist queries improved with repository pattern usage.
- Changed parameter passing from context to context.Context.
- Updated all database queries with new repository methods.
- Added database timeout contexts for all database operations.
- Implemented transaction management using repository pattern.
- Implemented tv shows season emojis.

+++ internal/tmdb/tv/type.go
@@ -24,7 +24,7 @@ type Season struct {
 type Episode struct {
 	Id            int64  `json:"id"`
 	AirDate       string `json:"air_date"`
-	SeasonNumber  int64  `json:"season_number"`
-	EpisodeNumber int64  `json:"episode_number"`
-	Runtime       int64  `json:"runtime"`
+	SeasonNumber  int32  `json:"season_number"`
+	EpisodeNumber int32  `json:"episode_number"`
+	Runtime       int32  `json:"runtime"`
 }
